sfc-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

boto3==1.35.78

httpx==0.28.0
tenacity==9.0.0
structlog==24.4.0


